app/requirements.txt
Authlib==1.4.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.0
docker==7.1.0
Flask==3.0.3
flask-oidc==2.2.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauth2client==4.1.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.2.1
Werkzeug==3.0.3
